refactor(feedback): Renaming the helper file to be more representative of the functions inside of it.
